+++ 404.php
@@ -1,5 +1,5 @@
 <?php get_header(); ?>
-
+<div class="wrapper">
 	<main role="main">
 		<!-- section -->
 		<section>
@@ -9,8 +9,13 @@
 
 				<h1><?php _e( 'Page not found', 'alda' ); ?></h1>
 				<h2>
-					<a href="<?php echo home_url(); ?>"><?php _e( 'Return home?', 'alda' ); ?></a>
+					<a href="<?php echo home_url(); ?>"><?php _e( 'Return home', 'alda' ); ?></a>
 				</h2>
+				<h2>or</h2>
+				<form class="search" method="get" action="<?php echo home_url(); ?>" role="search">
+	<input class="search-input" type="search" name="s" placeholder="I want to find..."">
+	<button class="search-submit" type="submit" role="button"><?php _e( 'Search', 'alda' ); ?></button>
+</form>
 
 			</article>
 			<!-- /article -->
@@ -18,7 +23,6 @@
 		</section>
 		<!-- /section -->
 	</main>
-
-<?php get_sidebar(); ?>
+</div>
 
 <?php get_footer(); ?>