+++ archive.php
@@ -1,5 +1,5 @@
 <?php get_header(); ?>
-
+<div class="wrapper">
 	<main role="main">
 		<!-- section -->
 		<section>
@@ -13,7 +13,6 @@
 		</section>
 		<!-- /section -->
 	</main>
-
-<?php get_sidebar(); ?>
+</div>
 
 <?php get_footer(); ?>