+++ author.php
@@ -1,5 +1,5 @@
 <?php get_header(); ?>
-
+<div class="wrapper">
 	<main role="main">
 		<!-- section -->
 		<section>
@@ -7,52 +7,16 @@
 		<?php if (have_posts()): the_post(); ?>
 
 			<h1><?php _e( 'Author Archives for ', 'alda' ); echo get_the_author(); ?></h1>
+			
+			<?php rewind_posts();?>
 
-		<?php if ( get_the_author_meta('description')) : ?>
+			<?php get_template_part('loop'); ?>
 
-		<?php echo get_avatar(get_the_author_meta('user_email')); ?>
-
-			<h2><?php _e( 'About ', 'alda' ); echo get_the_author() ; ?></h2>
-
-			<?php echo wpautop( get_the_author_meta('description') ); ?>
-
-		<?php endif; ?>
-
-		<?php rewind_posts(); while (have_posts()) : the_post(); ?>
-
-			<!-- article -->
-			<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
-
-				<!-- post thumbnail -->
-				<?php if ( has_post_thumbnail()) : // Check if Thumbnail exists ?>
-					<a href="<?php the_permalink(); ?>" title="<?php the_title(); ?>">
-						<?php the_post_thumbnail(array(120,120)); // Declare pixel size you need inside the array ?>
-					</a>
-				<?php endif; ?>
-				<!-- /post thumbnail -->
-
-				<!-- post title -->
-				<h2>
-					<a href="<?php the_permalink(); ?>" title="<?php the_title(); ?>"><?php the_title(); ?></a>
-				</h2>
-				<!-- /Post title -->
-
-				<!-- post details -->
-				<span class="date"><?php the_time('F j, Y'); ?> <?php the_time('g:i a'); ?></span>
-				<span class="author"><?php _e( 'Published by', 'alda' ); ?> <?php the_author_posts_link(); ?></span>
-				<span class="comments"><?php comments_popup_link( __( 'Leave your thoughts', 'alda' ), __( '1 Comment', 'alda' ), __( '% Comments', 'alda' )); ?></span>
-				<!-- /post details -->
-
-				<?php html5wp_excerpt('html5wp_index'); // Build your custom callback length in functions.php ?>
-
-				<br class="clear">
-
-				<?php edit_post_link(); ?>
+			<?php get_template_part('pagination'); ?>
 
 			</article>
 			<!-- /article -->
 
-		<?php endwhile; ?>
 
 		<?php else: ?>
 
@@ -71,7 +35,5 @@
 		</section>
 		<!-- /section -->
 	</main>
-
-<?php get_sidebar(); ?>
-
+</div>
 <?php get_footer(); ?>