+++ category.php
@@ -1,10 +1,10 @@
 <?php get_header(); ?>
-
+<div class="wrapper">
 	<main role="main">
 		<!-- section -->
 		<section>
 
-			<h1><?php _e( 'Categories for ', 'alda' ); single_cat_title(); ?></h1>
+			<h1><?php single_cat_title(); ?></h1>
 
 			<?php get_template_part('loop'); ?>
 
@@ -17,5 +17,5 @@
 					<?php if(!function_exists('dynamic_sidebar') || !dynamic_sidebar('widget-area-3')) ?>
 				</div>
 	</main>
-
+</div>
 <?php get_footer(); ?>