+++ single.php
@@ -1,5 +1,5 @@
 <?php get_header(); ?>
-
+<div class="wrapper">
 	<main role="main">
 	<!-- section -->
 	<section>
@@ -9,37 +9,15 @@
 		<!-- article -->
 		<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
 
-			<!-- post thumbnail -->
-			<?php if ( has_post_thumbnail()) : // Check if Thumbnail exists ?>
-				<a href="<?php the_permalink(); ?>" title="<?php the_title(); ?>">
-					<?php the_post_thumbnail(); // Fullsize image for the single post ?>
-				</a>
-			<?php endif; ?>
-			<!-- /post thumbnail -->
-
 			<!-- post title -->
-			<h1>
-				<a href="<?php the_permalink(); ?>" title="<?php the_title(); ?>"><?php the_title(); ?></a>
-			</h1>
+			<h1><?php the_title(); ?></h1>
 			<!-- /post title -->
-
-			<!-- post details -->
-			<span class="date"><?php the_time('F j, Y'); ?> <?php the_time('g:i a'); ?></span>
-			<span class="author"><?php _e( 'Published by', 'alda' ); ?> <?php the_author_posts_link(); ?></span>
-			<span class="comments"><?php if (comments_open( get_the_ID() ) ) comments_popup_link( __( 'Leave your thoughts', 'alda' ), __( '1 Comment', 'alda' ), __( '% Comments', 'alda' )); ?></span>
-			<!-- /post details -->
 
 			<?php the_content(); // Dynamic Content ?>
 
 			<?php the_tags( __( 'Tags: ', 'alda' ), ', ', '<br>'); // Separated by commas with a line break at the end ?>
 
 			<p><?php _e( 'Categorised in: ', 'alda' ); the_category(', '); // Separated by commas ?></p>
-
-			<p><?php _e( 'This post was written by ', 'alda' ); the_author(); ?></p>
-
-			<?php edit_post_link(); // Always handy to have Edit Post Links available ?>
-
-			<?php comments_template(); ?>
 
 		</article>
 		<!-- /article -->
@@ -61,5 +39,5 @@
 	</section>
 	<!-- /section -->
 	</main>
-
+</div>
 <?php get_footer(); ?>