+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 ### Alda (mkmeese.com) Changelog
 
-##1.0.2
+##1.0.3 (February 5, 2017)
+* Removed sidebar from page, 404, archive
+* Added portfolio, timerow and content for timeline to front-page.php
+* Stylesheet in header, not in functions
+* Header image as css, not in Customize
+* Styling header elements
+* Moved wrapper div to get full screen header and footer
+* Added Noto Sans and Noto Serif to stylesheet
+* Styling footer elements
+* Styling Portfolio timeline
+* Added wrapper to 404, category, archive, page, single
+* Shortened header and remove logo if not on front page
+* Removed sidebar from page, 404, archive
+* Added search to 404.php
+* Modified single.php and loop.php
+* Modified author.php to use the loop
+
+##1.0.2 (February 4, 2017)
 * Added in second menu to header using widget area instead of nav 2
 * Added site title and tagline, gave them empty classes
 * Added header image, gave empty class
@@ -12,14 +29,12 @@
 * Added widget area 3 category.php and removed sidebar
 * Removed sidebar from single.php
 
-##1.0.1
-
+##1.0.1 (February 3, 2017)
 * Added custom Header images, defined default
 * Removed localization support
 * Added three column responsive grid
 
 ## 1.0.0 (February 2, 2017)
-
 * Initial commit - HTML5 Blank
 * Modifying HTML5 Blank to Alda including text domain (revisit functions, custom post and template demo)
 * Added Noto Sans and Noto Serif to header